refactor(fileinput): move FileInput to its folder with theme interface
#443 #234 #456
