+++ src/docs/pages/CardPage.tsx
@@ -1,5 +1,6 @@
 import type { FC } from 'react';
-import { Button, Card, Checkbox, Dropdown, Label, TextInput } from '../../lib';
+import { Button, Card, Dropdown, Label, TextInput } from '../../lib';
+import { Checkbox } from '../../lib/components/Checkbox';
 import type { CodeExample } from './DemoPage';
 import { DemoPage } from './DemoPage';
