+++ src/docs/pages/ModalPage.tsx
@@ -1,7 +1,8 @@
 import type { FC } from 'react';
 import { useState } from 'react';
 import { HiOutlineExclamationCircle } from 'react-icons/hi';
-import { Button, Checkbox, Label, Modal, Select, TextInput } from '../../lib';
+import { Button, Label, Modal, Select, TextInput } from '../../lib';
+import { Checkbox } from '../../lib/components/Checkbox';
 import type { CodeExample } from './DemoPage';
 import { DemoPage } from './DemoPage';
