+++ src/lib/components/FormControls/index.ts
@@ -1,5 +1,3 @@
-export * from '../Checkbox';
-export * from './FileInput';
 export * from './HelperText';
 export * from './Label';
 export * from './Radio';
