+++ src/lib/components/Modal/Modal.stories.tsx
@@ -2,7 +2,8 @@ import { action } from '@storybook/addon-actions';
 import type { Meta, Story } from '@storybook/react/types-6-0';
 import { HiOutlineExclamationCircle } from 'react-icons/hi';
 import { Button } from '../Button';
-import { Checkbox, Label, TextInput } from '../FormControls';
+import { Checkbox } from '../Checkbox';
+import { Label, TextInput } from '../FormControls';
 import type { ModalProps } from './Modal';
 import { Modal } from './Modal';
